requirements.txt
Django==2.2.3
gunicorn==19.9.0
djangorestframework==3.9.4